creating styles
