+++ styles/ExercisesListStyles.tsx
@@ -1,0 +1,19 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  row: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "space-around",
+    paddingVertical: 10,
+    borderBottomWidth: 1,
+    borderBottomColor: "gray",
+  },
+  header: {
+    flex: 1,
+    fontWeight: "bold",
+  },
+  item: {
+    flex: 1,
+  },
+});
